Former-commit-id: a1ef21f641a3ea39667d8f855d3c3d43b76117da

+++ insti/src/emcchain.h
@@ -13,7 +13,7 @@
 #include	<cstdio>
 #include	<iostream>
 
-typedef double fast;
+typedef float fast;
 typedef unsigned uint;
 
 // used in estimate_EMC() for now
